Implementing clients list

+++ app/src/main/java/com/mithril/mobilegoldenleaf/models/User.kt
@@ -4,7 +4,7 @@ import androidx.room.PrimaryKey
 
 abstract class User {
     @PrimaryKey(autoGenerate = true)
-    var id: Int = 0
+    var id: Long = 0L
     var name: String? = null
     var phoneNumber: String? = null
 
@@ -17,7 +17,7 @@ abstract class User {
         this.phoneNumber = phoneNumber
     }
 
-    constructor(id: Int, name: String, phoneNumber: String) {
+    constructor(id: Long, name: String, phoneNumber: String) {
         this.id = id
         this.name = name
         this.phoneNumber = phoneNumber
